# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SMRITI-OS - Sovereign Retail OS",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/items/": {
            "post": {
                "tags": [
                    "item-master"
                ],
                "summary": "Create Item",
                "description": "Create a new item and its initial price/stock matrix.\nEnsures item_code uniqueness per store.",
                "operationId": "create_item_api_v1_items__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "item-master"
                ],
                "summary": "List Items",
                "description": "List items from Shoper9 Itemmaster with pagination and search.\nIncludes total stock across all locations.",
                "operationId": "list_items_api_v1_items__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search by code or name",
                            "title": "Search"
                        },
                        "description": "Search by code or name"
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Department Id"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Is Active"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{stockno}/stock-matrix": {
            "get": {
                "tags": [
                    "item-master"
                ],
                "summary": "Get Stock Matrix",
                "description": "Returns the location-wise stock for a specific item from Shoper9.",
                "operationId": "get_stock_matrix_api_v1_items__stockno__stock_matrix_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "stockno",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Stockno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{item_id}/price": {
            "patch": {
                "tags": [
                    "item-master"
                ],
                "summary": "Update Price Level",
                "description": "Update a price level (MRP, Wholesale, etc.).\nPreserves history by setting valid_to on the previous price.",
                "operationId": "update_price_level_api_v1_items__item_id__price_patch",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Item Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PriceLevelUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/batch": {
            "post": {
                "tags": [
                    "item-master"
                ],
                "summary": "Batch Create Items",
                "description": "Bulk create items. \nParity with Shoper9 Excel Import: Omits duplicates if requested.",
                "operationId": "batch_create_items_api_v1_items_batch_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemBatchCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BatchCreateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BatchCreateResponse": {
                "properties": {
                    "success_count": {
                        "type": "integer",
                        "title": "Success Count"
                    },
                    "skipped_count": {
                        "type": "integer",
                        "title": "Skipped Count"
                    },
                    "error_count": {
                        "type": "integer",
                        "title": "Error Count"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ItemResponse"
                        },
                        "type": "array",
                        "title": "Items",
                        "default": []
                    },
                    "skipped_codes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Skipped Codes",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "success_count",
                    "skipped_count",
                    "error_count"
                ],
                "title": "BatchCreateResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ItemBatchCreate": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ItemCreate"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "omit_duplicates": {
                        "type": "boolean",
                        "title": "Omit Duplicates",
                        "default": true
                    },
                    "apply_common_data": {
                        "type": "boolean",
                        "title": "Apply Common Data",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "items"
                ],
                "title": "ItemBatchCreate"
            },
            "ItemCreate": {
                "properties": {
                    "item_code": {
                        "type": "string",
                        "maxLength": 20,
                        "title": "Item Code"
                    },
                    "item_name": {
                        "type": "string",
                        "maxLength": 40,
                        "title": "Item Name"
                    },
                    "department_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Department Id"
                    },
                    "brand": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Brand"
                    },
                    "supplier_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Supplier Id"
                    },
                    "size_group_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Size Group Id"
                    },
                    "colour": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Colour"
                    },
                    "colour_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Colour Code"
                    },
                    "mrp_paise": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Mrp Paise"
                    },
                    "cost_paise": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "exclusiveMinimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cost Paise"
                    },
                    "gst_rate": {
                        "type": "integer",
                        "title": "Gst Rate",
                        "description": "Must be 0, 5, 12, 18, or 28"
                    },
                    "hsn_code": {
                        "type": "string",
                        "title": "Hsn Code"
                    },
                    "shoper_recid": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shoper Recid"
                    },
                    "anal_codes": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Anal Codes"
                    },
                    "stock_matrix": {
                        "items": {
                            "$ref": "#/components/schemas/StockMatrixEntry"
                        },
                        "type": "array",
                        "title": "Stock Matrix",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "item_code",
                    "item_name",
                    "department_id",
                    "mrp_paise",
                    "gst_rate",
                    "hsn_code"
                ],
                "title": "ItemCreate"
            },
            "ItemResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "item_code": {
                        "type": "string",
                        "title": "Item Code"
                    },
                    "item_name": {
                        "type": "string",
                        "title": "Item Name"
                    },
                    "department_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Department Id"
                    },
                    "brand": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Brand"
                    },
                    "mrp_paise": {
                        "type": "integer",
                        "title": "Mrp Paise"
                    },
                    "gst_rate": {
                        "type": "integer",
                        "title": "Gst Rate"
                    },
                    "hsn_code": {
                        "type": "string",
                        "title": "Hsn Code"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "total_stock": {
                        "type": "integer",
                        "title": "Total Stock",
                        "default": 0
                    },
                    "shoper_recid": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shoper Recid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "item_code",
                    "item_name",
                    "department_id",
                    "mrp_paise",
                    "gst_rate",
                    "hsn_code",
                    "is_active",
                    "created_at"
                ],
                "title": "ItemResponse"
            },
...
            "PriceLevelUpdate": {
                "properties": {
                    "price_level": {
                        "type": "string",
                        "pattern": "^(mrp|wholesale|staff|custom_1|custom_2)$",
                        "title": "Price Level"
                    },
                    "price_paise": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Price Paise"
                    },
                    "valid_from": {
                        "type": "string",
                        "format": "date",
                        "title": "Valid From"
                    }
                },
                "type": "object",
                "required": [
                    "price_level",
                    "price_paise"
                ],
                "title": "PriceLevelUpdate"
            },
...
            "StockMatrixEntry": {
                "properties": {
                    "size": {
                        "type": "string",
                        "title": "Size"
                    },
                    "colour": {
                        "type": "string",
                        "title": "Colour"
                    },
                    "qty_on_hand": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Qty On Hand",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "size",
                    "colour"
                ],
                "title": "StockMatrixEntry"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}